pyproject.toml
[tool.poetry]
name = "simple-airflow-automation"
version = "0.1.0"
description = "Automation of web scraping with Apache Airflow."
authors = ["Antony Nyagah <tony.m.nyagah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "^2.9.3"
Scrapy = "^2.11.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
